+++ src/nanollama/data/gssm.py
@@ -22,7 +22,7 @@ import torch
 from numpy.random import Generator, SeedSequence, default_rng
 from scipy.stats import dirichlet
 
-from ..cluster import get_rank
+from ..distributed import get_rank
 
 logger = logging.getLogger(__name__)
