+++ src/nanollama/utils.py
@@ -9,8 +9,36 @@ located in the root directory of this repository.
 @ 2025, Meta
 """
 
-from dataclasses import fields, is_dataclass
+from dataclasses import dataclass, fields, is_dataclass
 from typing import Any, TypeVar
+
+from torch.distributed.checkpoint.stateful import Stateful
+
+from .data import DataLoaderState
+from .optim import OptimizerState
+
+# -------------------------------------------------------------------------------
+# Training state
+# -------------------------------------------------------------------------------
+
+
+@dataclass
+class TrainState(Stateful):
+    data: DataLoaderState
+    optim: OptimizerState
+
+    def state_dict(self) -> dict[str, Any]:
+        return {"data": self.data.state_dict(), "optim": self.optim.state_dict()}
+
+    def load_state_dict(self, state_dict: dict[str, Any]) -> None:
+        self.data.load_state_dict(state_dict["data"])
+        self.optim.load_state_dict(state_dict["optim"])
+
+
+# -------------------------------------------------------------------------------
+# Initialization of nested configuration classes
+# -------------------------------------------------------------------------------
+
 
 T = TypeVar("T")
